requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
cfgv==3.3.1
colorama==0.4.5
decorator==5.1.1
distlib==0.3.6
Django==4.1.1
django-countries==7.3.2
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
executing==1.0.0
Faker==14.2.0
filelock==3.8.0
identify==2.5.4
inflection==0.5.1
ipython==8.5.0
jedi==0.18.1
jsonschema==4.15.0
matplotlib-inline==0.1.6
nodeenv==1.7.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.31
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
virtualenv==20.16.4
wcwidth==0.2.5
gunicorn==20.1.0
django-environ==0.9.0